# api/OpenAPI.yml
openapi: '3.0.2'
info:
  title: AIFEX OpenAPI
  version: '0.10.11'
servers:
  - url: http://localhost/api
paths:
...
  /evaluator/{sessionId}:
    patch:
        description: patch evaluator of a session
        parameters:
          - in: path
            name: sessionId
            schema:
              type: string
            required: true
            description: Id of the session
        requestBody:
          content:
            application/json:
              schema:
                type: object
                properties:
                  description:
                    type: string
                  expression:
                    type: string
        responses:
         '200':
            description: The evaluator of the session
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    sessionId:
                      type: string
         '403':
            description: Unauthorized
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Message'
         '404':
            description: NotFound
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Message'
        security:
          - bearerAuth: []
  /evaluator/evaluate:
    post:
        description: evaluate an expression
        parameters:
          - in: path
            name: sessionId
            schema:
              type: string
            required: true
            description: Id of the session
        requestBody:
          content:
            application/json:
              schema:
                type: object
                properties:
                  description:
                    type: string
                  expression:
                    type: string
        responses:
         '200':
            description: The evaluator of the session
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    sessionId:
                      type: string
         '403':
            description: Unauthorized
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Message'
         '404':
            description: NotFound
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Message'
        security:
          - bearerAuth: []
  /evaluator/expressionToDot:
    post:
      description: get a dotfile from an expression
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                expression:
                  type: string   
      responses:
        '200':
          description: returns a dotfile content
          content:
            application/json:
              schema:
                type: object
                properties:
                  expressionIsValid:
                    type: boolean   
                  dot:
                    type: string
        '400':
          description: Invalid parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '403':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
...
components:
  schemas:
    Message:
      type: object
      properties:
        message:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT   